added float64 so my output isnt 0 in the division sequence

+++ main.go
@@ -23,7 +23,7 @@ func main () {
 	fmt.Println("2009       *     2027           =   " + fmt.Sprint(2009 * 2027))
 
 	// division
-	fmt.Println("2009       /     2027           =   " + fmt.Sprint(2009 / 2027))
+	fmt.Println("2009       /     2027           =   " + fmt.Sprint(float64(2009) / 2027))
 
 	// Modulo
 	fmt.Println("2009       %     2027           =   " + fmt.Sprint(2009 % 2027))
